Download larger amounts of genbank data into R

This function works the same as rentrez::entrez_search() and ape::read.GenBank() together to download data from Genbank. However, it overcomes issues of download limits by iteratively downloading data and adding sequences and associated metadata together.

Usage

DoreyGenbank(
  db = "nucleotide",
  term = "(cytb[Gene Name]) AND (Apis[Organism])",
  retmax = 500,
  config = NULL,
  retmode = "xml",
  use_history = FALSE,
  seq.names = NULL,
  species.names = TRUE,
  as.character = FALSE,
  chunk.size = 400,
  quiet = TRUE,
  type = "DNA"
)

Arguments

db

character, name of the database to search for.

term

character, the search term. The syntax used in making these searches is described in the Details of this help message, the package vignette and reference given below.

retmax

numeric. The maximum number of sequences to return and download in a single search.

config

vector configuration options passed to httr::GET

retmode

character, one of json (default) or xml. This will make no difference in most cases.

use_history

logical. If TRUE return a web_history object for use in later calls to the NCBI

seq.names

the names to give to each sequence; by default the accession numbers are used. CURRENTLY NOT IN USE

species.names

a logical indicating whether to attribute the species names to the returned object.

as.character

a logical controlling whether to return the sequences as an object of class "DNAbin" (the default).

chunk.size

the number of sequences downloaded together (see details).

quiet

a logical value indicating whether to show the progress of the downloads. If TRUE, will also print the (full) name of the FASTA file containing the downloaded sequences.

type

a character specifying to download "DNA" (nucleotide) or "AA" (amino acid) sequences.

Value

A list of DNA sequences made of vectors of class "DNAbin", or of single characters (if as.character = TRUE) with two attributes (species and description).

Examples

TEST <- DoreyGenbank(db = "nucleotide",
term = "(cytb[Gene Name]) AND (Apis[Organism])",
retmax = 500)
#> Warning: URL 'https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db=nucleotide&id=1316031996,3033591882,2735580613,3021221636,2674565784,2674565782,2674565780,2674565778,2674565776,2674565774,2674565772,2674565770,2674565768,2674565766,2674565764,2674565762,2674565760,2674565758,2674565756,2674565754,2674565752,2674565750,2674565748,2674565746,2674565744,2674565742,2674565740,2674565738,2674565736,2674565734,2674565732,2674565730,2674565728,2674565726,2674565724,2674565722,2674565720,2674565718,2674565716,2674565714,2674565712,2674565710,2674565708,2674565706,2674565704,2674565702,2674565700,2674565698,2674565696,2674565694,2674565692,2674565690,2674565688,2674565686,2674565684,2674565682,2674565680,2674565678,2674565676,2674565674,2674565672,2674565670,2674565668,2674565666,2674565664,2674565662,2674565660,2674565658,2674565656,2674565654,2674565652,2674565650,2674565648,2674565646,2674565644,2674565642,2674565640,2674565638,2674565636,2674565634,2674565632,2674565630,2674565628,2674565 [... truncated]
#> Error in scan(file = URL, what = "", sep = "\n", quiet = TRUE, n = -1): cannot read from connection